simplify config parsing
Most of the changes are the consequence of using the `rc` module:

It is now possible to use .json or .ini format for config file.
`.tilesrc` is the default name of the config file -
`--config` option can now be used to specify config file location

All options - port, bind, max-age - can now be specified from either the
command line or the config file.